+++ nlp2fn/git/__init__.py
@@ -1,1 +1,0 @@
-from git.repo import Repo