+++ nlp2fn/sourcehandler/__init__.py
@@ -1,0 +1,1 @@
+from nlp2fn.sourcehandler.sourcemodel import SourceModel