+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.0.4 - 2023-05-26
+- Refactored code, Added Source Controller which is responsible for handling and keeping record of sources and related NLP statement.
+- Added Interface
+- NLP statement is searched inside all sources. Previously NLP statement related to a selected source was only used.
+- Create Proper Function to NLP record.
+- Provided option to updated added source.
+
+
 ## v0.0.3 - 2023-05-24
 
 - Added Check to make sure a directory is not added as function source for multiple time.